FAT-20589 add C745
