--- cypress/e2e/requests/new-request-user-without-barcode.cy.js
@@ -1,123 +1,0 @@
-import { ITEM_STATUS_NAMES, REQUEST_LEVELS, REQUEST_TYPES } from '../../support/constants';
-import Permissions from '../../support/dictionary/permissions';
-import InventoryInstances from '../../support/fragments/inventory/inventoryInstances';
-import NewRequest from '../../support/fragments/requests/newRequest';
-import RequestDetail from '../../support/fragments/requests/requestDetail';
-import Requests from '../../support/fragments/requests/requests';
-import RequestsSearchResultsPane from '../../support/fragments/requests/requestsSearchResultsPane';
-import Location from '../../support/fragments/settings/tenant/locations/newLocation';
-import ServicePoints from '../../support/fragments/settings/tenant/servicePoints/servicePoints';
-import TopMenu from '../../support/fragments/topMenu';
-import UserEdit from '../../support/fragments/users/userEdit';
-import Users from '../../support/fragments/users/users';
-import generateItemBarcode from '../../support/utils/generateItemBarcode';
-import getRandomPostfix from '../../support/utils/stringTools';
-
-describe('Choose requester without a barcode', () => {
-  let userData = {};
-  const servicePoint = ServicePoints.getDefaultServicePointWithPickUpLocation();
-  const itemData = {
-    barcode: generateItemBarcode(),
-    instanceTitle: `AT_C10956_Instance_${getRandomPostfix()}`,
-  };
-  let requestId;
-  let defaultLocation;
-
-  before('Prepare test data', () => {
-    cy.getAdminToken()
-      .then(() => {
-        ServicePoints.createViaApi(servicePoint);
-        defaultLocation = Location.getDefaultLocation(servicePoint.id);
-        Location.createViaApi(defaultLocation);
-        cy.getInstanceTypes({ limit: 1 }).then((instanceTypes) => {
-          itemData.instanceTypeId = instanceTypes[0].id;
-        });
-        cy.getHoldingTypes({ limit: 1 }).then((res) => {
-          itemData.holdingTypeId = res[0].id;
-        });
-        cy.getLoanTypes({ limit: 1 }).then((res) => {
-          itemData.loanTypeId = res[0].id;
-        });
-        cy.getDefaultMaterialType().then((res) => {
-          itemData.materialTypeId = res.id;
-          itemData.materialTypeName = res.name;
-        });
-      })
-      .then(() => {
-        InventoryInstances.createFolioInstanceViaApi({
-          instance: {
-            instanceTypeId: itemData.instanceTypeId,
-            title: itemData.instanceTitle,
-          },
-          holdings: [
-            {
-              holdingsTypeId: itemData.holdingTypeId,
-              permanentLocationId: defaultLocation.id,
-            },
-          ],
-          items: [
-            {
-              barcode: itemData.barcode,
-              status: { name: ITEM_STATUS_NAMES.AVAILABLE },
-              permanentLoanType: { id: itemData.loanTypeId },
-              materialType: { id: itemData.materialTypeId },
-            },
-          ],
-        });
-      })
-      .then(() => {
-        cy.createTempUser([Permissions.uiRequestsAll.gui], null, 'staff', false)
-          .then((userProperties) => {
-            userData = userProperties;
-          })
-          .then(() => {
-            UserEdit.addServicePointsViaApi([servicePoint.id], userData.userId, servicePoint.id);
-
-            cy.login(userData.username, userData.password, {
-              path: TopMenu.requestsPath,
-              waiter: RequestsSearchResultsPane.waitLoading,
-            });
-          });
-      });
-  });
-
-  after('Delete test data', () => {
-    cy.getAdminToken();
-    Requests.deleteRequestViaApi(requestId);
-    InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(itemData.barcode);
-    UserEdit.changeServicePointPreferenceViaApi(userData.userId, [servicePoint.id]);
-    Users.deleteViaApi(userData.userId);
-    Location.deleteInstitutionCampusLibraryLocationViaApi(
-      defaultLocation.institutionId,
-      defaultLocation.campusId,
-      defaultLocation.libraryId,
-      defaultLocation.id,
-    );
-    ServicePoints.deleteViaApi(servicePoint.id);
-  });
-
-  it(
-    'C10956 Choose requester without a barcode (vega)',
-    { tags: ['extendedPath', 'vega', 'C10956'] },
-    () => {
-      NewRequest.openNewRequestPane();
-      NewRequest.enterItemInfo(itemData.barcode);
-      NewRequest.verifyItemInformation([userData.barcode, ITEM_STATUS_NAMES.CHECKED_OUT]);
-      NewRequest.addRequester(userData.username);
-      NewRequest.verifyRequesterInformationWithBarcode(userData.username);
-      NewRequest.chooseRequestType(REQUEST_TYPES.PAGE);
-      NewRequest.choosePickupServicePoint(servicePoint.name);
-      NewRequest.saveRequestAndClose();
-      cy.wait('@createRequest').then((intercept) => {
-        requestId = intercept.response.body.id;
-        // Request is created
-        cy.location('pathname').should('eq', `/requests/view/${requestId}`);
-      });
-      RequestDetail.checkRequestInformation({
-        type: REQUEST_TYPES.PAGE,
-        status: 'Open - Not yet filled',
-        level: REQUEST_LEVELS.ITEM,
-      });
-    },
-  );
-});
